# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Fitnessstracker server.",
        "title": "Fitnesstracker API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/forget-password": {
            "post": {
                "description": "after otp verification set new password with Email, EventType, Token, NewePassword",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "set new password for user in case of forgot password",
                "operationId": "user-forgotpassword",
                "parameters": [
                    {
                        "description": "The input for set new password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ForgotPasswordInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password successfully changed",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data ",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "401": {
                        "description": "Invalid email or eventType or token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "login user with Email, Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "login for a user",
                "operationId": "user-login",
                "parameters": [
                    {
                        "description": "The input for login for user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserLogin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "401": {
                        "description": "Email id doesn't exist / Wrong password",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/otp/request": {
            "post": {
                "description": "send otp in registered email for set new user password in case of forgot password with Email, EventType",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "otp for forgot user password",
                "operationId": "user-otprequest",
                "parameters": [
                    {
                        "description": "The input for otp for forgot password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RequestOTP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OTP sent to email Successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "401": {
                        "description": "Email id doesn't exist",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/otp/verify": {
            "post": {
                "description": "otp verification for otp sent in registered email for set new user password in case of forgot password with Email, EventType, OTP",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "verify otp for forgot user password",
                "operationId": "user-verifyotp",
                "parameters": [
                    {
                        "description": "The input for verify otp for forgot password",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VerifyOTP"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Token"
                        }
                    },
                    "401": {
                        "description": "Invalid OTP",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "410": {
                        "description": "OTP Expired",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "add new user details with Email, FullName, Password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Add a new user",
                "operationId": "user-signup",
                "parameters": [
                    {
                        "description": "The input for add new user",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserSignup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserID"
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "409": {
                        "description": "This record contains duplicated data that conflicts with what is already in the database",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "498": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.Message"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ForgotPasswordInput": {
            "type": "object",
            "required": [
                "email",
                "eventType",
                "newPassword",
                "token"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string",
                    "enum": [
                        "forgot_password"
                    ]
                },
                "newPassword": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.Message": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.RequestOTP": {
            "type": "object",
            "required": [
                "email",
                "eventType"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string",
                    "enum": [
                        "forgot_password"
                    ]
                }
            }
        },
        "models.Token": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.UserID": {
            "type": "object",
            "required": [
                "userId"
            ],
            "properties": {
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.UserLogin": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.UserSignup": {
            "type": "object",
            "required": [
                "email",
                "fullName",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "models.VerifyOTP": {
            "type": "object",
            "required": [
                "email",
                "eventType",
                "otp"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "eventType": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "UserIDAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}